some refactoring due to new data type BigInteger

+++ bonaparte-android/src/main/java/de/jpaw/bonaparte/android/LinearLayoutComposer.java
@@ -1,6 +1,7 @@
 package de.jpaw.bonaparte.android;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.text.DecimalFormat;
 import java.util.HashMap;
 import java.util.Map;
@@ -205,7 +206,7 @@ abstract public class LinearLayoutComposer  implements MessageComposer<RuntimeEx
 
     // int(n)
     @Override
-    public void addField(BasicNumericElementaryDataItem di, Integer n) {
+    public void addField(BasicNumericElementaryDataItem di, BigInteger n) {
         if (n != null) {
             newTextView(di, Integer.toString(n));
         } else {
